refactor: move guilds.py to server/guilds/endpoints.py
Reorganized guild-related endpoints to follow the same structure as other
server endpoints (settings, logs, reminders, etc.).

Changes:
- Moved routers/v2/guilds.py to routers/v2/server/guilds/endpoints.py
- Created routers/v2/server/guilds/__init__.py
- Updated import path in startup.py

This maintains consistency with the rest of the server endpoint structure.
